Renamed main.ts
- Renamed main.ts to index.ts.
